Added fetching to use cases through the domain layer;

+++ src/main/java/com/shipping/demo/domain/tracking/TrackingDto.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @Setter
 @ToString(callSuper = true)
 public class TrackingDto {
-    private Integer parcelId;
+    private Long parcelId;
     private LocalDateTime eventDate;
-    private Short eventId;
+    private Long eventId;
 }
